done manage page

+++ SelfManagement/src/main/java/quynh/java/sm/langlearning/english/controller/VideoGroupAPIController.java
@@ -10,8 +10,10 @@ import javax.servlet.http.HttpServletResponse;
 
 import com.google.gson.Gson;
 
+import quynh.java.sm.langlearning.english.model.Video;
 import quynh.java.sm.langlearning.english.model.VideoGroup;
 import quynh.java.sm.langlearning.english.service.VideoGroupService;
+import quynh.java.sm.support.app.message.ResponseMessage;
 import quynh.java.sm.support.app.message.SMAction;
 import quynh.java.sm.support.app.message.SMError;
 import quynh.java.sm.support.app.message.SMMessage;
@@ -23,6 +25,7 @@ import quynh.java.sm.support.app.message.SMStatus;
 public class VideoGroupAPIController extends HttpServlet {
 	private static final long serialVersionUID = 1L;
     private VideoGroupService videoGroupService;
+    private ResponseMessage responseMessage;
     private Gson gson;
     /**
      * @see HttpServlet#HttpServlet()
@@ -30,6 +33,7 @@ public class VideoGroupAPIController extends HttpServlet {
     public VideoGroupAPIController() {
         super();
         videoGroupService = new VideoGroupService();
+        responseMessage = new ResponseMessage();
         gson = new Gson();
     }
 
@@ -59,10 +63,11 @@ public class VideoGroupAPIController extends HttpServlet {
 	private SMMessage addVideoGroup(HttpServletRequest request) {
 		int userId = 1;
 		String groupName = request.getParameter("groupName");
-		int resultAdd = videoGroupService.addVideoGroup(new VideoGroup(0, userId, groupName));
+		VideoGroup videoGroup = videoGroupService.addVideoGroup(new VideoGroup(0, userId, groupName));
 		SMMessage mes = new SMMessage();
-		if (resultAdd == 1) {
+		if (videoGroup != null) {
 			mes.setStatus(SMStatus.SUCCESS.getStatus());
+			mes.setData(videoGroup);
 		}
 		else {
 			mes.setStatus(SMStatus.FAIL.getStatus());
@@ -73,14 +78,16 @@ public class VideoGroupAPIController extends HttpServlet {
 	private SMMessage deleteVideoGroup(HttpServletRequest request) {
 		int userId = 1;
 		String groupName = request.getParameter("groupName");
-		int resultDel = videoGroupService.deleteVideoGroup(new VideoGroup(0, userId, groupName));
-		SMMessage mes = new SMMessage();
-		if (resultDel == 1) {
-			mes.setStatus(SMStatus.SUCCESS.getStatus());
-		}
-		else {
-			mes.setStatus(SMStatus.FAIL.getStatus());
-			mes.setData(SMError.UNKOWN_ERROR.getStatus());
+		SMMessage mes = null;
+		if (groupName == null || groupName.trim().isEmpty()) {
+			mes = responseMessage.createResponseMessage(SMStatus.FAIL, null);
+		} else {
+			int resultDel = videoGroupService.deleteVideoGroup(new VideoGroup(0, userId, groupName));
+			if (resultDel == 1) 
+				mes = responseMessage.createResponseMessage(SMStatus.SUCCESS, null);
+			else {
+				responseMessage.createResponseMessage(SMStatus.FAIL, SMError.UNKOWN_ERROR.getStatus());
+			}
 		}
 		return mes;
 	}
